Conflicts:
	logs/log4j2-demo.log
	src/main/java/com/mjvs/jgsp/controller/PassengerController.java
	src/main/java/com/mjvs/jgsp/controller/ZoneController.java
	src/main/java/com/mjvs/jgsp/controller/method_helpers/ResponseHelpers.java
	src/main/java/com/mjvs/jgsp/dto/StopDTO.java
	src/main/java/com/mjvs/jgsp/dto/ZoneDTO.java
	src/main/java/com/mjvs/jgsp/service/StopService.java
	src/main/java/com/mjvs/jgsp/service/StopServiceImpl.java
	src/main/resources/log4j2.xml

+++ src/main/java/com/mjvs/jgsp/service/StopServiceImpl.java
@@ -7,8 +7,8 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.HashMap;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class StopServiceImpl implements StopService
@@ -21,13 +21,12 @@ public class StopServiceImpl implements StopService
     @Override
     public boolean add(Stop newStop)
     {
-        Optional<Stop> stop = stopRepository.findAll().stream()
-                .filter(s -> s.getLatitude() == newStop.getLatitude()
-                        && s.getLongitude() == newStop.getLongitude()).findFirst();
-        if(stop.isPresent())
+        Stop stop = stopRepository.findByLatitudeAndLongitude(
+                newStop.getLatitude(), newStop.getLongitude());
+        if(stop != null)
         {
-            logger.debug(String.format("Stop %s %f %f already exists.",
-                    newStop.getName(), newStop.getLatitude(), newStop.getLongitude()));
+            logger.debug(String.format("Stop %f, %f already exists.",
+                    newStop.getLatitude(), newStop.getLongitude()));
             return false;
         }
 
@@ -39,8 +38,8 @@ public class StopServiceImpl implements StopService
         }
         catch (Exception ex)
         {
-            logger.error(String.format("Error adding new stop %f,%f message %s",
-                    newStop.getLatitude(), newStop.getLongitude(), ex.getMessage()));
+            logger.error(String.format("Error adding new stop %s %f,%f message %s",
+                    newStop.getName(), newStop.getLatitude(), newStop.getLongitude(), ex.getMessage()));
             return false;
         }
 
@@ -48,8 +47,14 @@ public class StopServiceImpl implements StopService
     }
 
     @Override
-    public boolean delete(Stop stop)
+    public boolean delete(double latitude, double longitude)
     {
+        Stop stop = stopRepository.findByLatitudeAndLongitude(latitude, longitude);
+        if(stop == null){
+            logger.warn(String.format("Stop %f, %f does not exist.", latitude, longitude));
+            return false;
+        }
+
         try
         {
             stopRepository.delete(stop);
@@ -69,5 +74,97 @@ public class StopServiceImpl implements StopService
     @Override
     public List<Stop> getAll() {
         return stopRepository.findAll();
+    }
+
+    @Override
+    public boolean rename(HashMap<String, String> data)
+    {
+        String oldName;
+        String newName;
+        double longitude;
+        double latitude;
+        try
+        {
+            oldName = data.get("oldName");
+            newName = data.get("newName");
+            longitude = new Double(data.get("longitude"));
+            latitude = new Double(data.get("latitude"));
+        }
+        catch (Exception ex)
+        {
+            logger.error(String.format("Wrong data format! %s", ex.getMessage()));
+            return false;
+        }
+
+        Stop stop = stopRepository.findByLatitudeAndLongitude(latitude, longitude);
+        if(stop == null)
+        {
+            logger.warn(String.format("Stop %f, %f does not exist.", latitude, longitude));
+            return false;
+        }
+
+        stop.setName(newName);
+
+        try
+        {
+            stopRepository.save(stop);
+            logger.info(String.format("Stop %s successfully renamed to %s",
+                    oldName, newName));
+        }
+        catch (Exception ex)
+        {
+            logger.info(String.format("Error renaming stop %s to %s",
+                    oldName, newName));
+            return false;
+        }
+
+        return true;
+    }
+
+    @Override
+    public boolean changeCoordinates(HashMap<String, Double> data)
+    {
+        double oldLat;
+        double oldLng;
+        double newLat;
+        double newLng;
+
+        try
+        {
+            oldLat = data.get("oldLat");
+            oldLng = data.get("oldLng");
+            newLat = data.get("newLat");
+            newLng = data.get("newLng");
+        }
+        catch (Exception ex)
+        {
+            logger.error(String.format("Wrong data format! %s", ex.getMessage()));
+            return false;
+        }
+
+        Stop stop = stopRepository.findByLatitudeAndLongitude(oldLat, oldLng);
+        if(stop == null)
+        {
+            logger.warn(String.format("Stop %f, %f does not exist.", oldLat, oldLng));
+            return false;
+        }
+
+        stop.setLatitude(newLat);
+        stop.setLongitude(newLng);
+
+        try
+        {
+            stopRepository.save(stop);
+            logger.info(String.format("Stop %f, %f successfully changed to %f, %f",
+                    oldLat, oldLng, newLat, newLng));
+        }
+        catch (Exception ex)
+        {
+            logger.info(String.format("Error changing stop %f, %f to %f, %f",
+                    oldLat, oldLng, newLat, newLng));
+            return false;
+        }
+
+        return true;
     }
 }
